analyze_pdf/requirements.txt
azure-functions
azure-storage-blob
PyMuPDF
fitz
requests